Initial Handle Older Version Process Import
